+++ packages/react/src/DefaultIcon.js
@@ -1,0 +1,20 @@
+import React from 'react'
+import { Dynamic } from './'
+
+let iconMap = {
+  SortAscending: () => <span>▲</span>,
+  SortDescending: () => <span>▼</span>,
+  MoveLeft: () => <span>←</span>,
+  MoveRight: () => <span>→</span>,
+  RemoveColumn: () => <span>x</span>,
+  AddColumn: () => <span>+</span>,
+  FilterExpand: () => <span>></span>,
+  FilterCollapse: () => <span>V</span>,
+  FilterAdd: () => <span>+</span>,
+  TableColumnMenu: () => <span>:</span>,
+  FilterListExpand: () => <span>◀</span>,
+  FilterListCollapse: () => <span>▼</span>,
+}
+let DefaultIcon = ({ icon }) => <Dynamic component={iconMap[icon]} />
+
+export default DefaultIcon