+++ packages/react/src/FilterList.js
@@ -1,20 +1,30 @@
 import React from 'react'
 import _ from 'lodash/fp'
 import { observer, inject } from 'mobx-react'
-import { Dynamic, SpacedList } from './layout'
+import { Dynamic } from './layout'
 import InjectTreeNode from './utils/injectTreeNode'
+import DefaultIcon from './DefaultIcon'
 
 export let Label = inject(_.pick('tree'))(
-  observer(({ tree, node, ...x }) => (
-    <div style={{ margin: '10px 0' }}>
-      <b {...x} />
+  observer(({ tree, node, Icon, ...x }) => (
+    <div
+      className="filter-field-label"
+      style={{
+        cursor: 'pointer',
+        display: 'flex',
+        justifyContent: 'space-between',
+      }}
+      onClick={() =>
+        tree && node && tree.mutate(node.path, { paused: !node.paused })
+      }
+    >
+      <span {...x} />
       {tree &&
         node && (
-          <span
-            style={{ float: 'right', marginRight: '5px', cursor: 'pointer' }}
-            onClick={() => tree.mutate(node.path, { paused: !node.paused })}
-          >
-            {node.paused ? '◀' : '▼'}
+          <span className="filter-field-label-icon">
+            <Icon
+              icon={node.paused ? 'FilterListExpand' : 'FilterListCollapse'}
+            />
           </span>
         )}
     </div>
@@ -23,32 +33,48 @@ export let Label = inject(_.pick('tree'))(
 Label.displayName = 'Label'
 
 export let FieldLabel = InjectTreeNode(
-  observer(({ node, node: { field } = {}, fields }) => (
-    <Label node={node}>{_.get([field, 'label'], fields)}</Label>
+  observer(({ node, node: { field } = {}, fields, Icon, label }) => (
+    <Label node={node} Icon={Icon}>
+      {label || _.get([field, 'label'], fields)}
+    </Label>
   ))
 )
 FieldLabel.displayName = 'FieldLabel'
 
 export let FilterList = InjectTreeNode(
   observer(
-    ({ node, typeComponents: types, fields, mapNodeToProps = _.noop }) => (
-      <SpacedList>
+    ({
+      node,
+      typeComponents: types,
+      fields,
+      mapNodeToProps = _.noop,
+      mapNodeToLabel = _.noop,
+      Icon = DefaultIcon,
+    }) => (
+      <div>
         {_.map(
           child => (
-            <div key={child.path}>
-              <FieldLabel node={child} fields={fields} />
+            <div key={child.path} className="filter-list-item">
+              <FieldLabel
+                node={child}
+                fields={fields}
+                Icon={Icon}
+                label={mapNodeToLabel(child, fields, types)}
+              />
               {!child.paused && (
-                <Dynamic
-                  component={types[child.type]}
-                  path={child.path.slice()}
-                  {...mapNodeToProps(child, fields, types)}
-                />
+                <div className="filter-list-item-contents">
+                  <Dynamic
+                    component={types[child.type]}
+                    path={child.path.slice()}
+                    {...mapNodeToProps(child, fields, types)}
+                  />
+                </div>
               )}
             </div>
           ),
           node.children
         )}
-      </SpacedList>
+      </div>
     )
   )
 )